Search and Update patient information was implemented and checking text fields for null value is implemented before initiating object but still has logical errors

+++ Utils/Email.cs
@@ -36,7 +36,7 @@ namespace webTRON_Management_Software.Utils
             return isSucess;
         }
         //Method to send verification code
-        public static bool SendVerificationCode(string userEmail,int verificationCode)
+        public static bool SendVerificationCode(string userEmail,double verificationCode)
         {
             //Set default status false
             bool isSucess = false;
